refactoring: NavBar 컴포넌트
* NavBar 컴포넌트에서 AppBody 컴포넌트 분리
* useNavBar hook 분리
